Organize files for orbit_coadd creation
