+++ subtract_zodi.py
@@ -1,5 +1,4 @@
-from file_handler import WISEMap
-from fullskymapping import FullSkyMap
+from wise_images_2_orbit_coadd.file_handler import WISEMap
 
 
 def subtract_zodi():
