+++ test.py
@@ -1,4 +1,3 @@
-from fullskymapping import FullSkyMap
 import matplotlib.pyplot as plt
 import numpy as np
