+++ test_minimization.py
@@ -1,4 +1,4 @@
-from file_handler import ZodiMap, HealpixMap
+from wise_images_2_orbit_coadd.file_handler import ZodiMap, HealpixMap
 import numpy as np
 from scipy.optimize import minimize
 import pickle
