+++ wise_images_2_orbit_coadd/process_manager.py
